+++ test/framework/src/main/java/org/opensearch/snapshots/mockstore/MockRepository.java
@@ -139,6 +139,8 @@ public class MockRepository extends FsRepository {
 
     private volatile boolean blockOnDataFiles;
 
+    private volatile boolean blockOnSegmentFiles;
+
     private volatile boolean blockOnDeleteIndexN;
 
     /**
@@ -190,6 +192,7 @@ public class MockRepository extends FsRepository {
         maximumNumberOfFailures = metadata.settings().getAsLong("max_failure_number", 100L);
         blockOnAnyFiles = metadata.settings().getAsBoolean("block_on_control", false);
         blockOnDataFiles = metadata.settings().getAsBoolean("block_on_data", false);
+        blockOnSegmentFiles = metadata.settings().getAsBoolean("block_on_segment", false);
         blockAndFailOnWriteSnapFile = metadata.settings().getAsBoolean("block_on_snap", false);
         randomPrefix = metadata.settings().get("random", "default");
         waitAfterUnblock = metadata.settings().getAsLong("wait_after_unblock", 0L);
@@ -237,6 +240,7 @@ public class MockRepository extends FsRepository {
         blocked = false;
         // Clean blocking flags, so we wouldn't try to block again
         blockOnDataFiles = false;
+        blockOnSegmentFiles = false;
         blockOnAnyFiles = false;
         blockAndFailOnWriteIndexFile = false;
         blockOnWriteIndexFile = false;
@@ -257,6 +261,14 @@ public class MockRepository extends FsRepository {
 
     public void setBlockOnAnyFiles(boolean blocked) {
         blockOnAnyFiles = blocked;
+    }
+
+    public void blockOnSegmentFiles(boolean blocked) {
+        blockOnSegmentFiles = blocked;
+    }
+
+    public void setBlockOnSegmentFiles(boolean blocked) {
+        blockOnSegmentFiles = blocked;
     }
 
     public void setBlockAndFailOnWriteSnapFiles(boolean blocked) {
@@ -306,6 +318,7 @@ public class MockRepository extends FsRepository {
         boolean wasBlocked = false;
         try {
             while (blockOnDataFiles
+                || blockOnSegmentFiles
                 || blockOnAnyFiles
                 || blockAndFailOnWriteIndexFile
                 || blockOnWriteIndexFile
@@ -407,6 +420,8 @@ public class MockRepository extends FsRepository {
                         blockExecutionAndMaybeWait(blobName);
                     } else if (blobName.startsWith("snap-") && blockAndFailOnWriteSnapFile) {
                         blockExecutionAndFail(blobName);
+                    } else if (blockOnSegmentFiles && blobName.contains(".si__")) {
+                        blockExecutionAndMaybeWait(blobName);
                     }
                 }
             }